+++ src/client/symbols/MyopiaSymbol.tsx
@@ -9,6 +9,12 @@ import {
 } from 'react-icons/fa';
 import { ORIENTATIONS } from '../../data/primitives';
 import { RiCheckboxBlankFill } from 'react-icons/ri';
+import {
+  TbChevronDownLeft,
+  TbChevronDownRight,
+  TbChevronUpLeft,
+  TbChevronUpRight,
+} from 'react-icons/tb';
 
 export interface MyopiaProps {
   textClass: string;
@@ -43,28 +49,58 @@ export default memo(function MyopiaSymbol({ textClass, symbol }: MyopiaProps) {
           <FaLongArrowAltRight size={'0.5em'} />
         </div>
       )}
-      {symbol.directions['up-left'] && (
+      {symbol.directions['up-left'] ? (
         <div key="up-left" className="absolute top-[0.12em] left-[0.12em]">
           <FaLongArrowAltLeft size={'0.5em'} className="rotate-45" />
         </div>
+      ) : (
+        symbol.diagonals && (
+          <div key="up-left" className="absolute top-[0.12em] left-[0.12em]">
+            <TbChevronUpLeft size={'0.3em'} className="opacity-50" />
+          </div>
+        )
       )}
-      {symbol.directions['up-right'] && (
+      {symbol.directions['up-right'] ? (
         <div key="up-right" className="absolute top-[0.12em] right-[0.12em]">
           <FaLongArrowAltUp size={'0.5em'} className="rotate-45" />
         </div>
+      ) : (
+        symbol.diagonals && (
+          <div key="up-right" className="absolute top-[0.12em] right-[0.12em]">
+            <TbChevronUpRight size={'0.3em'} className="opacity-50" />
+          </div>
+        )
       )}
-      {symbol.directions['down-left'] && (
+      {symbol.directions['down-left'] ? (
         <div key="down-left" className="absolute bottom-[0.12em] left-[0.12em]">
           <FaLongArrowAltDown size={'0.5em'} className="rotate-45" />
         </div>
+      ) : (
+        symbol.diagonals && (
+          <div
+            key="down-left"
+            className="absolute bottom-[0.12em] left-[0.12em]"
+          >
+            <TbChevronDownLeft size={'0.3em'} className="opacity-50" />
+          </div>
+        )
       )}
-      {symbol.directions['down-right'] && (
+      {symbol.directions['down-right'] ? (
         <div
           key="down-right"
           className="absolute bottom-[0.12em] right-[0.12em]"
         >
           <FaLongArrowAltRight size={'0.5em'} className="rotate-45" />
         </div>
+      ) : (
+        symbol.diagonals && (
+          <div
+            key="down-right"
+            className="absolute bottom-[0.12em] right-[0.12em]"
+          >
+            <TbChevronDownRight size={'0.3em'} className="opacity-50" />
+          </div>
+        )
       )}
       {ORIENTATIONS.every(d => !symbol.directions[d]) && (
         <div key="right" className="absolute m-auto">